rename : Common 폴더 위치 변경
